x/paper/types/message_update_paper_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"paper/testutil/sample"
)

func TestMsgUpdatePaper_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdatePaper
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdatePaper{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdatePaper{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
